requirements.txt
bottle==0.12.10
python-magic==0.4.12
